(core) extend attachment support to native S3
Summary:
This takes Spoffy's work for MinIO and applies it to the native S3 library used by Grist Labs for its SaaS. The goal is to make external attachments operate in the Grist Labs staging environment. Production will not be affected until the environment variable required for external attachments is set there.

Spoffy will be reworking some methods this touches, so there'll be some updated needed then, but there's so little code involved it doesn't seem worth waiting, compared to the advantage of hammering a bit on external attachments in staging as we do for other features.

Test Plan: existing tests pass

Reviewers: Spoffy

Reviewed By: Spoffy

Differential Revision: https://phab.getgrist.com/D4461
